refactor: Add BotInitializer - Created interface UpdateProcessor in service package - UpdateProcessor renamed to UpdateProcessorImpl and moved to service.impl package - Bot initialization method extracted to separate class BotInitializer
