+++ dispatcher/src/main/java/com/andybzg/dispatcher/controller/WebHookController.java
@@ -1,5 +1,6 @@
 package com.andybzg.dispatcher.controller;
 
+import com.andybzg.dispatcher.service.UpdateProcessor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
