+++ dispatcher/src/main/java/com/andybzg/dispatcher/service/UpdateProcessor.java
@@ -1,0 +1,11 @@
+package com.andybzg.dispatcher.service;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface UpdateProcessor {
+
+    void processUpdate(Update update);
+
+    void setView(SendMessage sendMessage);
+}
